Revert "Add trademark landing page"
