--- app/components/containers/trademark-home.js
@@ -1,4 +1,0 @@
-// Internal dependencies
-import TrademarkHome from 'components/ui/trademark-home';
-
-export default TrademarkHome;
