+++ app/components/ui/button/styles.scss
@@ -9,7 +9,7 @@
 	cursor: pointer;
 	display: inline-block;
 	font-size: 1.6rem;
-	font-weight: 400;
+	font-weight: 600;
 	margin-left: auto;
 	margin-right: auto;
 	padding: 15px;
