+++ app/components/ui/sunrise-home/styles.scss
@@ -65,3 +65,81 @@
 	padding: 1em 0;
 	text-align: center;
 }
+
+.explanations-wrapper {
+	background: $white;
+	font-size: 1.5em;
+	padding: 3em 0;
+}
+
+.explanations {
+	color: $blue-dark;
+	margin: 0 auto;
+	max-width: 900px;
+
+	@include breakpoint( '>660px' ) {
+		display: flex;
+		justify-content: space-between;
+	}
+}
+
+.explanations-heading {
+	color: $blue-dark;
+	font-family: $explanation-font;
+	font-size: 3rem;
+	font-weight: 100;
+	text-align: center;
+}
+
+.explanations-heading-line {
+	border: 1px solid $blue-bright;
+	margin: 1em auto;
+	width: 100px;
+}
+
+$icon-radius: 40px;
+
+.explanation-block {
+	border: 1px solid $gray-light;
+	border-radius: 2px;
+	display: block;
+	margin: $icon-radius + 20px 0;
+	padding: $icon-radius + 20px 1em 1em;
+	position: relative;
+
+	@include breakpoint( '>660px' ) {
+		flex-basis: 28%;
+	}
+
+	&::before {
+		background: lighten( $blue-bright, 20% );
+		border-radius: 50%;
+		content: '';
+		display: block;
+		height: $icon-radius * 2;
+		left: calc( 50% - #{$icon-radius} );
+		position: absolute;
+		top: 0 - $icon-radius;
+		width: $icon-radius * 2;
+	}
+}
+
+.explanation-title {
+	display: block;
+	font-size: 1.6rem;
+	font-weight: bold;
+	margin-bottom: 10px;
+
+	@include breakpoint( '>660px' ) {
+		text-align: center;
+	}
+}
+
+.explanation-text {
+	margin-bottom: 1em;
+}
+
+.explanation-link {
+	color: $blue-bright;
+	text-decoration: none;
+}
