--- app/components/ui/trademark-home/styles.scss
@@ -1,76 +1,0 @@
-@import 'app/styles/breakpoints';
-@import 'app/styles/colors';
-@import 'app/styles/fonts';
-
-.header-container {
-	background: $gray-light;
-}
-
-.header {
-	margin: 0 auto;
-	max-width: 720px;
-	padding-bottom: 30px;
-	text-align: center;
-}
-
-.logo {
-	color: $blue;
-	display: inline-block;
-	font-size: 2em;
-	font-weight: 200;
-	margin: 20px 0 70px;
-	text-decoration: none;
-
-	em {
-		font-weight: 600;
-	}
-}
-
-.heading {
-	color: $blue-dark;
-	font-family: $heading-font;
-	font-size: 3.7rem;
-	font-weight: 400;
-}
-
-.sub-heading {
-	color: $blue-dark-sunrise;
-	font-size: 2.3rem;
-	font-weight: 200;
-	margin-top: 15px;
-}
-
-.domain-input-container {
-	margin-bottom: 50px;
-
-	@include breakpoint( '>660px' ) {
-		display: flex;
-		margin-top: 50px;
-	}
-}
-
-.domain-input {
-	border: solid 1px $gray-medium;
-
-	@include breakpoint( '<660px' ) {
-		margin: 15px;
-	}
-
-	@include breakpoint( '>660px' ) {
-		width: 60%;
-	}
-}
-
-.pre-register-domain {
-	background: $blue-dark-sunrise;
-	color: $gray-light;
-	font-size: 1.6rem;
-	font-weight: 200;
-	padding: 20px;
-	text-align: center;
-}
-
-.pre-register-link {
-	color: $blue-medium;
-	text-decoration: none;
-}
